improved structure of server directory
