+++ server/endpoints/CreatedCustomIndexEndpoint.js
