+++ server/endpoints/CustomInternalServerErrorEndpoint.js
