+++ server/events/OnStaticGeneratorsChangeEvent.js
