+++ server/events/OnTemplatesChangeEvent.js
