+++ server/run.js
@@ -7,15 +7,15 @@ const { IsMaster, ClusterWithForkedWorkers, ClusterWithExitEvent } = require('@c
 const { ParsedJSON, Value } = require('@cuties/json')
 const { Backend, RestApi, CreatedServingFilesEndpoint } = require('@cuties/rest')
 const { ReadDataByPath, WatcherWithEventTypeAndFilenameListener } = require('@cuties/fs')
-const CustomNotFoundEndpoint = require('./../CustomNotFoundEndpoint')
-const CustomInternalServerErrorEndpoint = require('./../CustomInternalServerErrorEndpoint')
-const CreatedCustomIndexEndpoint = require('./../CreatedCustomIndexEndpoint')
-const OnPageStaticJsFilesChangeEvent = require('./../OnPageStaticJsFilesChangeEvent')
-const OnStaticGeneratorsChangeEvent = require('./../OnStaticGeneratorsChangeEvent')
-const OnTemplatesChangeEvent = require('./../OnTemplatesChangeEvent')
-const ReloadedBackendOnFailedWorkerEvent = require('./../ReloadedBackendOnFailedWorkerEvent')
-const UrlToFSPathMapper = require('./../UrlToFSPathMapper')
-const PrintedToConsolePageLogo = require('./../PrintedToConsolePageLogo')
+const CustomNotFoundEndpoint = require('./endpoints/CustomNotFoundEndpoint')
+const CustomInternalServerErrorEndpoint = require('./endpoints/CustomInternalServerErrorEndpoint')
+const CreatedCustomIndexEndpoint = require('./endpoints/CreatedCustomIndexEndpoint')
+const OnPageStaticJsFilesChangeEvent = require('./events/OnPageStaticJsFilesChangeEvent')
+const OnStaticGeneratorsChangeEvent = require('./events/OnStaticGeneratorsChangeEvent')
+const OnTemplatesChangeEvent = require('./events/OnTemplatesChangeEvent')
+const ReloadedBackendOnFailedWorkerEvent = require('./events/ReloadedBackendOnFailedWorkerEvent')
+const PrintedToConsolePageLogo = require('./async/PrintedToConsolePageLogo')
+const UrlToFSPathMapper = require('./async/UrlToFSPathMapper')
 
 const numCPUs = require('os').cpus().length
 const env = process.env.NODE_ENV || 'local'
