+++ test/server/async/PrintedToConsolePageLogo.js
@@ -1,0 +1,23 @@
+'use strict'
+
+/*
+ logo in file ./static/txt/logo.txt is generated by
+  http://patorjk.com/software/taag/#p=display&h=0&v=0&f=Calvin%20S&t=%20P%20A%20G%20E%0A
+*/
+
+const { AsyncObject } = require('@cuties/cutie')
+
+class PrintedToConsolePageLogo extends AsyncObject {
+  constructor (logo, version, step) {
+    super(logo, version, step)
+  }
+
+  syncCall () {
+    return (logo, version, step) => {
+      console.log(`${logo}${version}\n\n${step}\n`)
+      return logo
+    }
+  }
+}
+
+module.exports = PrintedToConsolePageLogo
