+++ test/server/endpoints/CreatedCustomIndexEndpointTest.js
@@ -1,10 +1,10 @@
 'use strict'
 
 const { IndexEndpoint } = require('@cuties/rest')
-const CreatedCustomIndexEndpoint = require('./../../server/CreatedCustomIndexEndpoint')
-const CustomNotFoundEndpoint = require('./../../server/CustomNotFoundEndpoint')
 const { Assertion } = require('@cuties/assert')
 const { Is } = require('@cuties/is')
+const CreatedCustomIndexEndpoint = require('./../../../server/endpoints/CreatedCustomIndexEndpoint')
+const CustomNotFoundEndpoint = require('./../../../server/endpoints/CustomNotFoundEndpoint')
 
 new Assertion(
   new Is(
