+++ test/server/endpoints/CustomIndexEndpointTest.js
@@ -1,11 +1,11 @@
 'use strict'
 
 const { IndexEndpoint } = require('@cuties/rest')
-const CustomIndexEndpoint = require('./../../server/CustomIndexEndpoint')
-const CustomNotFoundEndpoint = require('./../../server/CustomNotFoundEndpoint')
 const { Assertion } = require('@cuties/assert')
 const { Is } = require('@cuties/is')
-const { CustomStream, InvokedEndpoint } = require('./common')
+const { CustomStream, InvokedEndpoint } = require('./../common')
+const CustomIndexEndpoint = require('./../../../server/endpoints/CustomIndexEndpoint')
+const CustomNotFoundEndpoint = require('./../../../server/endpoints/CustomNotFoundEndpoint')
 
 new Assertion(
   new Is(
