+++ test/server/endpoints/CustomInternalServerErrorEndpointTest.js
@@ -1,10 +1,10 @@
 'use strict'
 
 const { InternalServerErrorEndpoint } = require('@cuties/rest')
-const CustomInternalServerErrorEndpoint = require('./../../server/CustomInternalServerErrorEndpoint')
 const { Assertion } = require('@cuties/assert')
 const { Is } = require('@cuties/is')
-const { CustomStream, InvokedEndpoint } = require('./common')
+const { CustomStream, InvokedEndpoint } = require('./../common')
+const CustomInternalServerErrorEndpoint = require('./../../../server/endpoints/CustomInternalServerErrorEndpoint')
 
 new Assertion(
   new Is(
