+++ test/server/events/NotFoundErrorEventTest.js
@@ -3,9 +3,9 @@
 const { AsyncObject } = require('@cuties/cutie')
 const { Assertion } = require('@cuties/assert')
 const { Is } = require('@cuties/is')
-const NotFoundErrorEvent = require('./../../server/NotFoundErrorEvent')
-const CustomNotFoundEndpoint = require('./../../server/CustomNotFoundEndpoint')
-const { CustomStream } = require('./common')
+const { CustomStream } = require('./../common')
+const NotFoundErrorEvent = require('./../../../server/events/NotFoundErrorEvent')
+const CustomNotFoundEndpoint = require('./../../../server/endpoints/CustomNotFoundEndpoint')
 
 class InvokedNotFoundErrorEvent extends AsyncObject {
   constructor (event) {
