+++ test/server/events/OnStaticGeneratorsChangeEventTest.js
@@ -1,11 +1,12 @@
 'use strict'
 
 const { AsyncObject } = require('@cuties/cutie')
-const OnTemplatesChangeEvent = require('./../../server/OnTemplatesChangeEvent')
 const { Assertion } = require('@cuties/assert')
 const { Is } = require('@cuties/is')
 
-class InvokedOnTemplatesChangeEvent extends AsyncObject {
+const OnStaticGeneratorsChangeEvent = require('./../../../server/events/OnStaticGeneratorsChangeEvent')
+
+class InvokedOnStaticGeneratorsChangeEvent extends AsyncObject {
   constructor (event, eventType, fileName) {
     super(event, eventType, fileName)
   }
@@ -20,14 +21,14 @@ class InvokedOnTemplatesChangeEvent extends AsyncObject {
 
 new Assertion(
   new Is(
-    new OnTemplatesChangeEvent('./test/server/files'),
+    new OnStaticGeneratorsChangeEvent('./test/server/files'),
     Function
   )
 ).after(
   new Assertion(
     new Is(
-      new InvokedOnTemplatesChangeEvent(
-        new OnTemplatesChangeEvent('./test/server/files'),
+      new InvokedOnStaticGeneratorsChangeEvent(
+        new OnStaticGeneratorsChangeEvent('./test/server/files'),
         'change', 'index.html'
       ),
       Function
@@ -35,8 +36,8 @@ new Assertion(
   ).after(
     new Assertion(
       new Is(
-        new InvokedOnTemplatesChangeEvent(
-          new OnTemplatesChangeEvent('./test/server/files'),
+        new InvokedOnStaticGeneratorsChangeEvent(
+          new OnStaticGeneratorsChangeEvent('./test/server/files'),
           'create', 'index.html'
         ),
         Function
